+++ source/matcher/h2_memcmp.cpp
@@ -1,12 +1,12 @@
-static inline h2_string readable_size(int width, int nbits)
+static inline h2_string readable_size(size_t width, size_t nbits)
 {
    char t[64];
    switch (width) {
-   case 1: sprintf(t, "%d bit%s", nbits, nbits > 1 ? "s" : ""); break;
-   case 8: sprintf(t, "%d byte%s", nbits / 8, nbits / 8 > 1 ? "s" : ""); break;
-   case 16: sprintf(t, "%d word%s", nbits / 16, nbits / 16 > 1 ? "s" : ""); break;
-   case 32: sprintf(t, "%d dword%s", nbits / 32, nbits / 32 > 1 ? "s" : ""); break;
-   case 64: sprintf(t, "%d qword%s", nbits / 64, nbits / 64 > 1 ? "s" : ""); break;
+   case 1: sprintf(t, "%d bit%s", (int)nbits, nbits > 1 ? "s" : ""); break;
+   case 8: sprintf(t, "%d byte%s", (int)(nbits / 8), nbits / 8 > 1 ? "s" : ""); break;
+   case 16: sprintf(t, "%d word%s", (int)(nbits / 16), nbits / 16 > 1 ? "s" : ""); break;
+   case 32: sprintf(t, "%d dword%s", (int)(nbits / 32), nbits / 32 > 1 ? "s" : ""); break;
+   case 64: sprintf(t, "%d qword%s", (int)(nbits / 64), nbits / 64 > 1 ? "s" : ""); break;
    default: sprintf(t, "?"); break;
    }
    return h2_string(t);
@@ -15,11 +15,11 @@ static inline h2_string readable_size(int width, int nbits)
 h2_inline h2_fail* h2_matches_bytecmp::matches(const void* a, int n, bool caseless, bool dont) const
 {
    bool result = false;
-   int _nbytes;
+   size_t _nbytes;
    if (isstring) {
       if (is_hex_string((const char*)e)) {
          unsigned char* _e = (unsigned char*)alloca(strlen((const char*)e));
-         int max_length = hex_to_bytes((const char*)e, _e);
+         size_t max_length = hex_to_bytes((const char*)e, _e);
          _nbytes = nbytes;
          if (nbytes == 0) _nbytes = max_length;
          if (_nbytes <= max_length) {
@@ -48,7 +48,7 @@ h2_inline h2_sentence h2_matches_bytecmp::expection(bool caseless, bool dont) co
 
 h2_inline h2_fail* h2_matches_bitcmp::matches(const void* a, int n, bool caseless, bool dont) const
 {
-   int max_length = INT_MAX;
+   size_t max_length = INT_MAX;
    unsigned char* _e = (unsigned char*)e;
    if (isstring) {
       unsigned char* t = (unsigned char*)alloca(strlen((const char*)e));
@@ -66,7 +66,7 @@ h2_inline h2_fail* h2_matches_bitcmp::matches(const void* a, int n, bool caseles
          return h2_fail::new_normal("length required");
       }
    }
-   int _nbits = nbits;
+   size_t _nbits = nbits;
    if (nbits == 0) _nbits = max_length;
    if (max_length < _nbits) {
       return h2_fail::new_normal("length too loog");
