+++ source/mock/h2_checkin.cpp
@@ -1,4 +1,4 @@
-h2_inline h2_fail* h2_checkin::check(const char* func, int index, int total, const char* file, int line)
+h2_inline h2_fail* h2_checkin::check(const char* func, size_t index, size_t total, const char* file, int line)
 {
    if (is_satisfied() || is_saturated()) return nullptr;
    h2_sentence t = func + gray("()") + " expected " + delta(expect(), "green") + " but actually " + delta(actual(), "red,bold") + " called";
