+++ source/mock/h2_checkin.hpp
@@ -10,7 +10,7 @@ struct h2_checkin { /* 考勤 ; 函数被调次数期望 */
    bool is_saturated(/*饱和*/) const { return call == most; }
    bool is_excessive(/*过多*/) const { return most < call; }
 
-   h2_fail* check(const char* func, int index, int total, const char* file, int line);
+   h2_fail* check(const char* func, size_t index, size_t total, const char* file, int line);
    const char* actual();
    const char* expect();
